Rewrite NN wrapper and combine 1B into one directory
