--- python/complete-2-B/config/data-config_mask.yaml
@@ -1,40 +1,0 @@
-new_variables:
-  pf_mask: awkward.ones_like(Part_E)
-  not_mask: 1 - Part_mask
-
-preprocess:
-  method: manual
-  data_fraction: 1.0
-
-inputs:
-  pf_points:
-    length: 52
-    vars:
-      - [Part_vx, null, 1, -10, 10]
-      - [Part_vy, null, 1, -10, 10]
-      - [Part_vz, null, 1, -10, 10]
-    pad_mode: constant
-    pad_value: 0
-
-  pf_features:
-    length: 52
-    vars:
-      - Part_E
-      - Part_px
-      - Part_py
-      - Part_pz
-    pad_mode: constant
-    pad_value: 0
-
-  pf_mask:
-    length: 52
-    vars:
-      - pf_mask
-
-labels:
-  type: simple
-  value: [Part_mask]
-  var_funcs: {}
-
-observers:
-  - pdg_id
