--- python/incomplete-1-B/config/data-config.yaml
@@ -1,39 +1,0 @@
-new_variables:
-  pf_mask: awkward.ones_like(Part_E)
-
-preprocess:
-  method: manual
-  data_fraction: 1.0
-
-inputs:
-  pf_points:
-    length: 35
-    vars:
-      - [Part_vx, null, 1, -10, 10]
-      - [Part_vy, null, 1, -10, 10]
-      - [Part_vz, null, 1, -10, 10]
-    pad_mode: constant
-    pad_value: 0
-
-  pf_features:
-    length: 35
-    vars:
-      - Part_E
-      - Part_px
-      - Part_py
-      - Part_pz
-    pad_mode: constant
-    pad_value: 0
-
-  pf_mask:
-    length: 35
-    vars:
-      - pf_mask
-
-labels:
-  type: simple
-  value: [B_E, B_px, B_py, B_pz]
-  var_funcs: {}
-
-observers:
-  - pdg_id
